requirements.txt
joblib
pandas==0.24.2
scipy==1.2.2
scikit-learn==0.20.4
google-cloud-storage==1.26.0
gcsfs
category_encoders
pygeohash
xgboost==0.90
mlflow
memoized_property
psutil
kaggle